+++ src/contacts/contacts.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+
+export default function Contacts() {
+  return (
+      <p>Contacts</p>
+  )}
