+++ src/header/header.css
@@ -63,4 +63,6 @@ a:focus:after {
 .top-menu ul li a:hover{
     font-weight: 700;
 }
-
+.active{
+    
+}
